# improve jsonmodel speed + some extra cleanup