--- a/src/base/types/JSONModel.jl
+++ b/src/base/types/JSONModel.jl
@@ -24,6 +24,12 @@ reactions(model) # see the list of reactions
 """
 struct JSONModel <: MetabolicModel
     json::Dict{String,Any}
+    rxn_index::Dict{String,Int}
+    rxns::Dict{String,Any}
+    met_index::Dict{String,Int}
+    mets::Dict{String,Any}
+    gene_index::Dict{String,Int}
+    genes::Dict{String,Any}
 end
 
 
@@ -49,13 +55,27 @@ macro _json_section(namesConstant::Symbol, error)
     ))
 end
 
-macro _json_section_firstid(namesConstant::Symbol, id::Symbol, error)
-    return esc(:(
-        begin
-            _s = @_json_section $namesConstant ($error)
-            _s[findfirst(x -> x["id"] == $id, _s)]
-        end
-    ))
+_json_rxn_name(r, i) = string(get(r, "id", "rxn$i"))
+_json_met_name(m, i) = string(get(m, "id", "met$i"))
+_json_gene_name(g, i) = string(get(g, "id", "gene$i"))
+
+JSONModel(json::Dict{String,Any}) = begin
+    rs = @_json_section rxns throw(
+        DomainError(keys(model.json), "JSON model has no reaction keys"),
+    )
+    ms = @_json_section mets throw(
+        DomainError(keys(model.json), "JSON model has no metabolite keys"),
+    )
+    gs = @_json_section genes []
+    JSONModel(
+        json,
+        Dict(i => _json_rxn_name(r, i) for (i, r) in enumerate(rs)),
+        rs,
+        Dict(i => _json_met_name(m, i) for (i, m) in enumerate(ms)),
+        ms,
+        Dict(i => _json_gene_name(g, i) for (i, g) in enumerate(gs)),
+        gs,
+    )
 end
 
 function _parse_annotations(x)::Annotations
@@ -73,37 +93,21 @@ _parse_notes(x)::Notes = _parse_annotations(x)
 
 Extract reaction names (stored as `.id`) from JSON model.
 """
-function reactions(model::JSONModel)
-    rs = @_json_section rxns throw(
-        DomainError(keys(model.json), "JSON model has no reaction keys"),
-    )
-
-    return [string(get(rs[i], "id", "rxn$i")) for i in eachindex(rs)]
-end
+reactions(model::JSONModel) = [_json_rxn_name(r, i) for (i, r) in enumerate(model.rxns)]
 
 """
     metabolites(model::JSONModel)
 
 Extract metabolite names (stored as `.id`) from JSON model.
 """
-function metabolites(model::JSONModel)
-    ms = @_json_section mets throw(
-        DomainError(keys(model.json), "JSON model has no metabolite keys"),
-    )
-
-    return [string(get(ms[i], "id", "met$i")) for i in eachindex(ms)]
-end
+metabolites(model::JSONModel) = [_json_met_name(m, i) for (i, m) in enumerate(model.mets)]
 
 """
     genes(model::JSONModel)
 
 Extract gene names from a JSON model.
 """
-function genes(model::JSONModel)
-    gs = @_json_section genes return []
-
-    return [string(get(gs[i], "id", "gene$i")) for i in eachindex(gs)]
-end
+genes(model::JSONModel) = [_json_gene_name(g, i) for (i, g) in enumerate(model.genes)]
 
 """
     stoichiometry(model::JSONModel)
@@ -115,16 +119,11 @@ function stoichiometry(model::JSONModel)
     rxn_ids = reactions(model)
     met_ids = metabolites(model)
 
-    rs = @_json_section rxns throw(
-        DomainError(keys(model.json), "JSON model has no reaction keys"),
-    )
-
     S = SparseArrays.spzeros(length(met_ids), length(rxn_ids))
     for (i, rid) in enumerate(rxn_ids)
-        r = rs[findfirst(x -> x["id"] == rid, rs)]
-        for (met_id, coeff) in r["metabolites"]
-            j = findfirst(==(met_id), met_ids)
-            if isnothing(j)
+        r = model.rxns[model.rxn_index[rid]]
+        for (mid, coeff) in r["metabolites"]
+            if !haskey(model.met_index, mid)
                 throw(
                     DomainError(
                         met_id,
@@ -132,7 +131,7 @@ function stoichiometry(model::JSONModel)
                     ),
                 )
             end
-            S[j, i] = coeff
+            S[model.met_index[mid], i] = coeff
         end
     end
     return S
@@ -144,148 +143,127 @@ end
 Get the bounds for reactions, assuming the information is stored in
 `.lower_bound` and `.upper_bound`.
 """
-function bounds(model::JSONModel)
-    rs = @_json_section rxns return (
-        sparse(fill(-_constants.default_reaction_bound, n_reactions(model))),
-        sparse(fill(_constants.default_reaction_bound, n_reactions(model))),
-    )
-    return (
-        sparse([get(rxn, "lower_bound", -_constants.default_reaction_bound) for rxn in rs]),
-        sparse([get(rxn, "upper_bound", _constants.default_reaction_bound) for rxn in rs]),
-    )
-end
+bounds(model::JSONModel) = (
+    sparse([
+        get(rxn, "lower_bound", -_constants.default_reaction_bound) for rxn in model.rxns
+    ]),
+    sparse([
+        get(rxn, "upper_bound", _constants.default_reaction_bound) for rxn in model.rxns
+    ]),
+)
 
 """
     objective(model::JSONModel)
 
 Collect `.objective_coefficient` keys from model reactions.
 """
-function objective(model::JSONModel)
-    rs = @_json_section rxns return spzeros(n_reactions(model))
-
-    return sparse([float(get(rxn, "objective_coefficient", 0.0)) for rxn in rs])
-end
+objective(model::JSONModel) =
+    sparse([float(get(rxn, "objective_coefficient", 0.0)) for rxn in model.rxns])
 
 """
     reaction_gene_associaton(model::JSONModel, rid::String)
 
 Parses the `.gene_reaction_rule` from reactions.
 """
-function reaction_gene_association(model::JSONModel, rid::String)
-    rxn = @_json_section_firstid rxns rid return nothing
-    return _maybemap(_parse_grr, get(rxn, "gene_reaction_rule", nothing))
-end
+reaction_gene_association(model::JSONModel, rid::String) = _maybemap(
+    _parse_grr,
+    get(model.rxns[model.rxn_index[rid]], "gene_reaction_rule", nothing),
+)
 
 """
     reaction_subsystem(model::JSONModel, rid::String)
 
 Parses the `.subsystem` out from reactions.
 """
-function reaction_subsystem(model::JSONModel, rid::String)
-    rxn = @_json_section_firstid rxns rid return nothing
-    return get(rxn, "subsystem", nothing)
-end
+reaction_subsystem(model::JSONModel, rid::String) =
+    get(model.rxns[model.rxn_index[rid]], "subsystem", nothing)
 
 """
     metabolite_formula(model::JSONModel, mid::String)
 
 Parse and return the metabolite `.formula`
 """
-function metabolite_formula(model::JSONModel, mid::String)
-    met = @_json_section_firstid mets mid return nothing
-    return _maybemap(_parse_formula, get(met, "formula", nothing))
-end
+metabolite_formula(model::JSONModel, mid::String) =
+    _maybemap(_parse_formula, get(model.mets[model.met_index[mid]], "formula", nothing))
 
 """
     metabolite_charge(model::JSONModel, mid::String)
 
 Return the metabolite `.charge`
 """
-function metabolite_charge(model::JSONModel, mid::String)
-    met = @_json_section_firstid mets mid return nothing
-    return get(met, "charge", 0)
-end
+metabolite_charge(model::JSONModel, mid::String) =
+    get(model.mets[model.met_index[mid]], "charge", 0)
 
 """
     metabolite_compartment(model::JSONModel, mid::String)
 
 Return the metabolite `.compartment`
 """
-function metabolite_compartment(model::JSONModel, mid::String)
-    met = @_json_section_firstid mets mid return nothing
-    return get(met, "compartment", nothing)
-end
+metabolite_compartment(model::JSONModel, mid::String) =
+    get(model.mets[model.met_index[mid]], "compartment", nothing)
 
 """
     gene_annotations(model::JSONModel, gid::String)::Annotations
 
 Gene annotations from the [`JSONModel`](@ref).
 """
-function gene_annotations(model::JSONModel, gid::String)::Annotations
-    gene = @_json_section_firstid genes gid return Dict()
-    _maybemap(_parse_annotations, get(gene, "annotation", nothing))
-end
+gene_annotations(model::JSONModel, gid::String)::Annotations = _maybemap(
+    _parse_annotations,
+    get(model.genes[model.gene_index[gid]], "annotation", nothing),
+)
 
 """
     gene_notes(model::JSONModel, gid::String)::Notes
 
 Gene notes from the [`JSONModel`](@ref).
 """
-function gene_notes(model::JSONModel, gid::String)::Notes
-    gene = @_json_section_firstid genes gid return Dict()
-    _maybemap(_parse_notes, get(gene, "notes", nothing))
-end
+gene_notes(model::JSONModel, gid::String)::Notes =
+    _maybemap(_parse_notes, get(model.genes[model.gene_index[gid]], "notes", nothing))
 
 """
     reaction_annotations(model::JSONModel, rid::String)::Annotations
 
 Reaction annotations from the [`JSONModel`](@ref).
 """
-function reaction_annotations(model::JSONModel, rid::String)::Annotations
-    rxn = @_json_section_firstid rxns rid return Dict()
-    _maybemap(_parse_annotations, get(rxn, "annotation", nothing))
-end
+reaction_annotations(model::JSONModel, rid::String)::Annotations = _maybemap(
+    _parse_annotations,
+    get(model.rxns[model.rxn_index[rid]], "annotation", nothing),
+)
 
 """
     reaction_notes(model::JSONModel, rid::String)::Notes
 
 Reaction notes from the [`JSONModel`](@ref).
 """
-function reaction_notes(model::JSONModel, rid::String)::Notes
-    rxn = @_json_section_firstid rxns rid return Dict()
-    _maybemap(_parse_notes, get(rxn, "notes", nothing))
-end
+reaction_notes(model::JSONModel, rid::String)::Notes =
+    _maybemap(_parse_notes, get(model.rxns[model.rxn_index[rid]], "notes", nothing))
 
 """
     metabolite_annotations(model::JSONModel, mid::String)::Annotations
 
 Metabolite annotations from the [`JSONModel`](@ref).
 """
-function metabolite_annotations(model::JSONModel, mid::String)::Annotations
-    met = @_json_section_firstid mets mid return Dict()
-    _maybemap(_parse_annotations, get(met, "annotation", nothing))
-end
+metabolite_annotations(model::JSONModel, mid::String)::Annotations = _maybemap(
+    _parse_annotations,
+    get(model.mets[model.met_index[mid]], "annotation", nothing),
+)
 
 """
     metabolite_notes(model::JSONModel, mid::String)::Notes
 
 Metabolite notes from the [`JSONModel`](@ref).
 """
-function metabolite_notes(model::JSONModel, mid::String)::Notes
-    met = @_json_section_firstid mets mid return Dict()
-    _maybemap(_parse_notes, get(met, "notes", nothing))
-end
+metabolite_notes(model::JSONModel, mid::String)::Notes =
+    _maybemap(_parse_notes, get(model.mets[model.met_index[mid]], "notes", nothing))
 
 """
-    reaction_stoichiometry(model::JSONModel, rxn_id::String)::Dict{String, Float64}
+    reaction_stoichiometry(model::JSONModel, rid::String)::Dict{String, Float64}
 
-Return the reaction equation of reaction with id `rxn_id` in model. The reaction
+Return the reaction equation of reaction with ID `rid` in model. The reaction
 equation maps metabolite ids to their stoichiometric coefficients.
 """
-function reaction_stoichiometry(m::JSONModel, rxn_id::String)::Dict{String,Float64}
-    ind = findfirst(x -> x["id"] == rxn_id, m.json["reactions"])
-    m.json["reactions"][ind]["metabolites"]
-end
+reaction_stoichiometry(model::JSONModel, rid::String)::Dict{String,Float64} =
+    model.rxns[model.rxn_index[rid]]["metabolites"]
 
 """
     Base.convert(::Type{JSONModel}, mm::MetabolicModel)