… src/base/types/StandardModel.jl
… src/base/types/abstract/MetabolicModel.jl
… src/reconstruction/CoreModel.jl
… src/reconstruction/CoreModelCoupled.jl