Fixes #350 by caching.