@stelmo 2021-07-26
Test code:
```julia
using COBREXA
using BenchmarkTools

@benchmark load_model(StandardModel, model_path)
```

Current conversion:
```julia
julia> @benchmark load_model(StandardModel, model_path)
BenchmarkTools.Trial: 
  memory estimate:  532.70 MiB
  allocs estimate:  30704399
  --------------
  minimum time:     7.763 s (2.35% GC)
  median time:      7.763 s (2.35% GC)
  mean time:        7.763 s (2.35% GC)
  maximum time:     7.763 s (2.35% GC)
  --------------
  samples:          1
  evals/sample:     1
```

This PR:
```julia
julia> @benchmark load_model(StandardModel, model_path)
BenchmarkTools.Trial: 
  memory estimate:  69.00 MiB
  allocs estimate:  1034802
  --------------
  minimum time:     223.278 ms (0.00% GC)
  median time:      256.775 ms (10.61% GC)
  mean time:        256.305 ms (11.40% GC)
  maximum time:     284.919 ms (17.28% GC)
  --------------
  samples:          20
  evals/sample:     1
```

I would say problem solved, thanks @exaexa 😄 🚀 👾 ✨ 🤖