@laurentheirendt 2021-07-27
the currently displayed status check from git-r3 was a misfire and can be ignored 👍🏼